requirements.txt
# Core
numpy>=1.24
scipy>=1.10
pyyaml>=6.0
tqdm>=4.66

# ML
torch>=2.1
# Optional / recommended (install matching wheels for your system):
# torchvision>=0.16
# torchaudio>=2.1

# Metrics & calibration
scikit-learn>=1.3

# Audio feature extraction (if torchaudio is unavailable)
librosa>=0.10

# Plotting (optional; used by scripts that make figures)
matplotlib>=3.7

# Model FLOPs (optional)
thop>=0.1.1.post2209072238
